+++ pymock_api/command/cmd_ps.py
@@ -27,7 +27,7 @@ from ..model.api_config import APIParameter as MockedAPIParameter
 from ..model.swagger_config import API as SwaggerAPI
 from ..model.swagger_config import APIParameter as SwaggerAPIParameter
 from ..server import BaseSGIServer, setup_asgi, setup_wsgi
-from .cmd import MockAPICommandParser, SubCommand
+from .options import MockAPICommandParser, SubCommand
 
 _COMMAND_CHAIN: List[Type["CommandProcessor"]] = []
